Input Validation
You can now set up input validation by giving input a validation method, normal, email of cellNumber or just leave it blank for no validation.

You then make an array of Inputs for each form, pass that to the Validate Input function and if the inputs are valid it will perform the function otherwise it won't and will scroll you to the incorrect input.